Convert identity into auth plugin
